Move unit tests to adapter.discover.test. Add a test that uses adapter.fetch
